fix(audio): scope stackchan-voice to CoreS3
